Fix: 리뷰 반영 - 1. 헤더 컴포넌트화 2. 스타일 수정
